requirements.txt
flask==2.3.3
google-generativeai==0.5.0
firebase-admin==6.5.0
Pillow==10.4.0
python-dotenv==1.0.0
gunicorn==21.2.0